Add Localization Provider and Update Dependency Injection
Commit Message:

Update the dependency injection setup to register the LanguageRepository and remove the unnecessary providers from the `di.dart` file.
Modify the `LoginView` and `MovieDetailsProvider` to use the new `LanguageProvider` and `MovieRepository` instances.
Refactor the `MainApp` to use the `LanguageProvider` and `LoginProvider` in a `MultiProvider` setup to provide the required dependencies to the app.