Improve `CustomElevatedButton` style
The commit modifies the `CustomElevatedButton` widget to apply bold font weight to the button text. Previously, the text was displayed in the default font weight, which may not have provided sufficient visual emphasis. By setting the `fontWeight` property to `FontWeight.bold`, the text will appear more prominent and aligned with the overall design of the application.